Test invite user.

+++ app/src/main/java/pt/ulisboa/tecnico/cmov/g15/airdesk/storage/FileSystemManager.java
@@ -194,7 +194,8 @@ public class FileSystemManager {
     }
 
     public static void deleteStorage() {
-        deleteRecursively(appDir);
+        if(appDir.exists())
+            deleteRecursively(appDir);
     }
 
 }
